Move session stuff to its own file
