+++ src/provider.h
@@ -264,8 +264,11 @@ void p11prov_uri_free(P11PROV_URI *parsed_uri);
 CK_OBJECT_CLASS p11prov_uri_get_class(P11PROV_URI *uri);
 CK_ATTRIBUTE p11prov_uri_get_id(P11PROV_URI *uri);
 char *p11prov_uri_get_object(P11PROV_URI *uri);
+char *p11prov_uri_get_pin(P11PROV_URI *uri);
+CK_RV p11prov_uri_match_token(P11PROV_URI *uri, CK_TOKEN_INFO *token);
 int p11prov_get_pin(const char *in, char **out);
 
+/* Sessions */
 P11PROV_SESSION *p11prov_session_new(P11PROV_CTX *ctx, CK_SLOT_ID slotid);
 P11PROV_SESSION *p11prov_session_ref(P11PROV_SESSION *session);
 CK_RV p11prov_session_open(P11PROV_SESSION *session, bool login,
